Break Qualifers out to top level class
