+++ robolectric-resources/src/main/java/org/robolectric/res/ResBundle.java
@@ -1,20 +1,15 @@
 package org.robolectric.res;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.StringTokenizer;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class ResBundle<T> {
-  // Matches a version qualifier like "v14". Parentheses capture the numeric
-  // part for easy retrieval with Matcher.group(2).
-  private static final Pattern VERSION_QUALIFIER_PATTERN = Pattern.compile("(v)([0-9]+)$");
-  private static final Pattern SIZE_QUALIFIER_PATTERN = Pattern.compile("(s?[wh])([0-9]+)dp$");
+
 
   private final ResMap<T> valuesMap = new ResMap<>();
   private final ResMap<List<T>> valuesArrayMap = new ResMap<>();
@@ -35,14 +30,6 @@ public class ResBundle<T> {
   public Value<T> getValue(ResName resName, String qualifiers) {
     List<Value<T>> values = valuesMap.find(maybeOverride(resName));
     return values != null ? pick(values, qualifiers) : null;
-  }
-
-  public static int getVersionQualifierApiLevel(String qualifiers) {
-    Matcher m = VERSION_QUALIFIER_PATTERN.matcher(qualifiers);
-    if (m.find()) {
-      return Integer.parseInt(m.group(2));
-    }
-    return -1;
   }
 
   public static <T> Value<T> pick(List<Value<T>> values, String qualifiersStr) {
@@ -183,99 +170,4 @@ public class ResBundle<T> {
     }
   }
 
-  private static class Qualifiers {
-    // Version are matched in the end, and hence have least order
-    private static final int ORDER_VERSION = 0;
-    // Various size qualifies, in increasing order of importance.
-    private static final List<String> INT_QUALIFIERS = Arrays.asList("v", "h", "w", "sh", "sw");
-    private static final int TOTAL_ORDER_COUNT = INT_QUALIFIERS.size();
-
-    private static Map<String, Qualifiers> sQualifiersCache = new HashMap<>();
-
-    private final int[] mWeights = new int[TOTAL_ORDER_COUNT];
-    // Set of all the qualifiers which need exact matching.
-    private final List<String> mDefaults = new ArrayList<>();
-
-    public boolean matches(Qualifiers other) {
-      if (!passesRequirements(other)) {
-        return false;
-      }
-      return other.mDefaults.containsAll(mDefaults);
-    }
-
-    public boolean passesRequirements(Qualifiers other) {
-      for (int i = 0; i < TOTAL_ORDER_COUNT; i++) {
-        if (other.mWeights[i] != -1 && mWeights[i] != -1 && other.mWeights[i] < mWeights[i]) {
-          return false;
-        }
-      }
-      return true;
-    }
-
-    public boolean isBetterThan(Qualifiers other, Qualifiers context) {
-      // Compare the defaults in the order they appear in the context.
-      for (String qualifier : context.mDefaults) {
-        if (other.mDefaults.contains(qualifier) ^ mDefaults.contains(qualifier)) {
-          return mDefaults.contains(qualifier);
-        }
-      }
-
-      for (int i = TOTAL_ORDER_COUNT -1 ; i > ORDER_VERSION; i--) {
-        if (other.mWeights[i] != mWeights[i]) {
-          return mWeights[i] > other.mWeights[i];
-        }
-      }
-
-      // Compare the version only if the context defines a version.
-      if (context.mWeights[ORDER_VERSION] != -1
-          && other.mWeights[ORDER_VERSION] != mWeights[ORDER_VERSION]) {
-        return mWeights[ORDER_VERSION] > other.mWeights[ORDER_VERSION];
-      }
-
-      // The qualifiers match completely
-      return false;
-    }
-
-    public static Qualifiers parse(String qualifiersStr) {
-      synchronized (sQualifiersCache) {
-        Qualifiers result = sQualifiersCache.get(qualifiersStr);
-        if (result != null) {
-          return result;
-        }
-        StringTokenizer st = new StringTokenizer(qualifiersStr, "-");
-        result = new Qualifiers();
-        // Version qualifiers are also allowed to match when only one of the qualifiers
-        // defines a version restriction.
-        result.mWeights[ORDER_VERSION] = -1;
-
-        while (st.hasMoreTokens()) {
-          String qualifier = st.nextToken();
-          if (qualifier.isEmpty()) {
-            continue;
-          }
-
-          Matcher m = VERSION_QUALIFIER_PATTERN.matcher(qualifier);
-          if (!m.find()) {
-            m = SIZE_QUALIFIER_PATTERN.matcher(qualifier);
-            if (!m.find()) {
-              m = null;
-            }
-          }
-          if (m != null) {
-            int order = INT_QUALIFIERS.indexOf(m.group(1));
-            if (order == ORDER_VERSION && result.mWeights[ORDER_VERSION] != -1) {
-              throw new IllegalStateException(
-                  "A resource file was found that had two API level qualifiers: " + qualifiersStr);
-            }
-            result.mWeights[order] = Integer.parseInt(m.group(2));
-          } else {
-            result.mDefaults.add(qualifier);
-          }
-        }
-
-        sQualifiersCache.put(qualifiersStr, result);
-        return result;
-      }
-    }
-  }
 }
